refactor(!meta/.../all_wrong)
BREAKING CHANGES

Made a module
